Added project transitions for New -> Unconfirmed -> Inactive -> Active -> Cancelled.

+++ app/controllers/projects_controller.rb
@@ -13,7 +13,7 @@ class ProjectsController < InheritedResources::Base
 
 
 	def index
-		@projects = Project.where("active=1 and confirmed = 1").order("end_date ASC").page(params[:page]).per(2)
+		@projects = Project.where(:active => 1, :confirmed => 1, :state => PROJ_STATES[2]).order("end_date ASC").page(params[:page]).per(2)
 		#@projects = Project.limit(9).where("active = 1 and confirmed = 1").order("end_date ASC")
 		#@projects1 = @projects.slice(0..2) || []
 		#@projects2 = @projects.slice(3..5) || []
@@ -25,7 +25,7 @@ class ProjectsController < InheritedResources::Base
 		@project = Project.new(params[:project])
 		@project.user_id = current_user.id
 		@project.payment_account_id = Project::UNDEFINED_PAYMENT_ACCOUNT_ID #To pass validation at valid?
-		@proejct.state = PROJ_STATS[0] #unconfirmed
+		@project.state = PROJ_STATES[0] #unconfirmed
 	
 		if @project.valid?
 			@project.save
@@ -37,19 +37,31 @@ class ProjectsController < InheritedResources::Base
 			render :action => :new
 		end
 	end
+	
+	def update
+		@project = Project.where(:name => params[:id]).first
+		if params[:activate] == 'true'
+			@project.state = PROJ_STATES[2] #active
+			@project.save!
+			flash[:notice] = "Successfully ACTIVATED project." 
+    elsif @project.update_attributes(params[:project])  
+      flash[:notice] = "Successfully updated project."  
+    end  
+    respond_with(@project)
+	end
 
 	def save
 		Amazon::FPS::AmazonLogger::log_recipient_token_response(params)
 
 		if !Amazon::FPS::AmazonValidator::valid_recipient_response?(save_project_url, session, params)
 			flash[:alert] = "An error occurred with your project. Please try again."
-			@proejct.state = PROJ_STATS[0] #unconfirmed
+			@proejct.state = PROJ_STATES[0] #unconfirmed
 			return redirect_to root_path
 		end
 
 		@project = Project.find_by_id(session[:project_id])
 		@project.confirmed = true
-		@proejct.state = PROJ_STATS[1] #inactive
+		@project.state = PROJ_STATES[1] #inactive
 		@project.payment_account_id = params[:tokenID]
 
 		session[:project_id] = nil
@@ -68,13 +80,23 @@ class ProjectsController < InheritedResources::Base
 	end
 
 	def destroy
-		if !@project.destroy
-			flash[:alert] = "Project could not be deleted. Please try again."
-			return redirect_to @project
+		if @project.state == PROJ_STATES[0] || @project.state == PROJ_STATES[1]
+			if !@project.destroy
+				flash[:alert] = "Project could not be deleted. Please try again."
+				return redirect_to @project
+			else 
+				flash[:alert] = "Project successfully deleted. Sorry to see you go!"
+				return redirect_to root_path
+			end
 		else 
-			flash[:alert] = "Project successfully deleted. Sorry to see you go!"
+			#project will not be deleted but will be CANCELED and only visible to user
+			@project.state = PROJ_STATES[5] #canceled
+			@project.save!
+			flash[:alert] = "Project successfully canceled. Project is now only visible to you."
 			return redirect_to root_path
 		end
+		flash[:alert] = "Somthing went wrong. Check out the project destory action in the controller."
+		return redirect_to root_path
 	end
 	
   def show
@@ -87,6 +109,10 @@ class ProjectsController < InheritedResources::Base
     
     @update = Update.new(params[:update])
     @updates = Update.where(:project_id => @project.id)
+  end
+  
+  def edit
+  	@project = Project.where(:name => params[:id]).first
   end
